Fix OSSMC NamespaceDetailsTab circular dependency crash
Move getNamespaceRevision helpers out of VirtualList/Renderers into
NamespaceRevisionUtils so NamespaceDetailsOverview no longer triggers
the Config ↔ Renderers TDZ when OSSMC lazy-loads NamespaceDetailsTab.
